requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
google-search-results>=2.4.2
